Split boost dependency out into separate file
Boost dependency is very complex and it is easier to manage it
in a separate file.
